Add SoapySDR gNB config
Using this config allows using LimeSDR and other SoapySDR supported
radios as gNBs. These radios do not support the clock and time source
settings functions in the UHD driver, hence they have those options
set to "unsupported".

Note: this configuration was tested on the LimeSDR Mini V2.6 but is
not stable due to hardware constraints and overheating. When a bag of
ice is placed on the radio, it can sustain a more stable connection
to a UE for a few minutes before overheating again and dropping.